worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "KinjiKawaguchi", email = "kawakin0310@icloud.com" }
]
dependencies = [
    "grpcio>=1.65.4",
    "google-api-python-client>=2.139.0",
    "protobuf>=5.27.3",
    "transformers>=4.43.3",
    "torch>=2.4.0",
    "manim>=0.18.1",
    "google-cloud-storage>=2.18.0",
]
readme = "../README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/worker"]
